backend/requirements.txt
fastapi[all]
uvicorn[standard]
motor
python-jose[cryptography]
bcrypt
python-dotenv

